Ensure ordered dict output even if PyYaml is modified
- PyYaml lib can be modified with representers.  Because we are sharing
PyYaml with multiple plugins, it is very likely that someone may force
ordered dict behavior to be different than we are forcing it.  Since I
don’t want to provide my own PyYaml lib for this plugin, and we only
need to ensure ordered output for the mail frontmatter, we will just
iterate the ordered dict outputting each key value pair as its own dict
and join each output to represent the entire ordered dict.  This will
ensure we are always handling things the way we want.